requirements.txt
bson==0.5.10
click==7.1.2
dnspython==1.16.0
flake8==3.8.4
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pymongo==3.11.0
python-dateutil==2.8.1
python-decouple==3.3
six==1.15.0
Werkzeug==1.0.1
